PD-2: Poprawa wyliczania średniej ilosci punktów

+++ src/main/java/pd2/Player.java
@@ -22,7 +22,7 @@ public class Player {
     }
 
     public double getAverage() {
-        double average = getSum() / scores.length;
+        double average = (double)getSum() / scores.length;
         return average;
     }
